Done backend version 1

+++ src/main/java/com/minhbui/ecommerce/service/ProductService.java
@@ -1,11 +1,15 @@
 package com.minhbui.ecommerce.service;
 
+import com.cloudinary.utils.ObjectUtils;
+import com.minhbui.ecommerce.config.CloudinaryConfig;
 import com.minhbui.ecommerce.dto.request.*;
+import com.minhbui.ecommerce.dto.response.ProductAttributesResponse;
 import com.minhbui.ecommerce.dto.response.ProductDetailResponse;
 import com.minhbui.ecommerce.dto.response.ProductResponse;
 import com.minhbui.ecommerce.dto.response.ProductSkusResponse;
 import com.minhbui.ecommerce.exception.AppCatchException;
 import com.minhbui.ecommerce.exception.ErrorCode;
+import com.minhbui.ecommerce.mapper.ProductAttributesMapper;
 import com.minhbui.ecommerce.mapper.ProductDetailMapper;
 import com.minhbui.ecommerce.mapper.ProductMapper;
 import com.minhbui.ecommerce.mapper.ProductskusMapper;
@@ -15,11 +19,20 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.Date;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -36,10 +49,17 @@ public class ProductService {
     EventRepository eventRepository;
     CategoryRepository categoryRepository;
     UserRepository userRepository;
+    ProductAttributesMapper productAttributesMapper;
+    ProductAttributesRepository productAttributesRepository;
+    UserService userService;
+    CloudinaryConfig cloudinary;
 
     //product
-
-    public ProductResponse craeteProductOfCategory(ProductCreationRequest request) {
+    @PreAuthorize("hasRole('SHOP')")
+    public ProductResponse craeteProductOfCategory(
+            ProductCreationRequest request,
+            List<MultipartFile> files
+    ) throws IOException {
 
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -60,7 +80,10 @@ public class ProductService {
             throw  new AppCatchException(ErrorCode.CATEGORY_DISABLE);
         }
 
+        Set<String> images = uploadMultiImg(files, String.valueOf(owner.getId()));
+
         Product product = productMapper.toProduct(request);
+        product.setImages(images);
 
         product.setShop(shop);
         product.setCategory(category);
@@ -77,8 +100,28 @@ public class ProductService {
         return productResponse;
     }
 
-    //lấy tất cả sản phẩm của shop sở hữu
-    public List<ProductResponse> getAllProductsOfShop() {
+    private Set<String> uploadMultiImg(List<MultipartFile> files, String unique) throws IOException {
+        Set<String> urlList = new HashSet<>();
+        for (MultipartFile file : files) {
+            var result = cloudinary.cloudinary().uploader()
+                    .upload(file.getBytes(), ObjectUtils.asMap(
+                            "folder", "ecommerce/shop_owner_"+unique+"/products"
+                    ));
+            urlList.add((String) result.get("secure_url"));;
+        }
+        return urlList;
+    }
+
+    //lấy tất cả sản phẩm
+    public Page<Product> getAllProducts(int page, int size) {
+        Pageable pageable = PageRequest.of(page,size);
+        return productRepository.findALlByDisabledFalse(pageable);
+    }
+
+    //lấy tất cả sản phẩm của shop sở hữu method for shop
+    @PreAuthorize("hasRole('SHOP')")
+    public List<ProductResponse> getAllProductsOfShop(int page, int size) {
+        Pageable pageable = PageRequest.of(page,size);
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
         User owner = userRepository.findByEmail(emailOwner)
@@ -87,23 +130,49 @@ public class ProductService {
         Shop shop = shopRepository.findByOwner(owner)
                 .orElseThrow(() ->new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER));
 
-        return productMapper.toProductResponseList(productRepository.findProductsByShopWithDisabledCategory(shop));
+        return productMapper.toProductResponseList(productRepository.findProductsByShopWithDisabledCategory(shop, pageable));
     }
 
-    //lấy tất cả sản phẩm của shop khi click vào shop
-    public List<ProductResponse> getAllProductsOfShopForUser(Long shopId) {
+    //tìm kieems
+    //tìm kiếm theo tên
+    public Page<Product> searchProductsByName(String name, Pageable pageable) {
+        return productRepository.findByNameContainingIgnoreCase(name, pageable);
+    }
+    //tìm kiếm theo tên và sắp xếp giá
+    public Page<Product> searchProductsByNameAndSortByPrice(String name, Pageable pageable) {
+        return productRepository.findByNameOrderByPriceAsc(name, pageable);
+    }
+    //tìm kiếm theo tên và khoản giá trong shop
+    public Page<Product> searchProductsByShopAndPriceRange(Long shopId, Long minPrice, Long maxPrice, Pageable pageable) {
+        Shop shop = shopRepository.findById(shopId)
+                .orElseThrow(() ->new AppCatchException(ErrorCode.SHOP_NOT_FOUND));
+        return productRepository.findByShopAndPriceRange(shop, minPrice, maxPrice, pageable);
+    }
+
+    //lấy tất cả sản phẩm của shop khi click vào shop method for user
+    public List<Product> getAllProductsOfShopForUser(Long shopId, int page, int size, boolean asc) {
+        Pageable pageable = PageRequest.of(page,size);
         Shop shop = shopRepository.findById(shopId)
                 .orElseThrow(() ->new AppCatchException(ErrorCode.SHOP_NOT_FOUND));
 
-        return productMapper.toProductResponseList(productRepository.findProductsByShopWithDisabledCategory(shop));
+        if(asc){
+            return productRepository.findByShopOrderByMinPriceAsc(shop, pageable);
+        }else {
+            return productRepository.findByShopOrderByMinPriceDesc(shop, pageable);
+        }
     }
 
     //lấy tất cả sản phẩm khi click vào category
-    public List<ProductResponse> getAllProductsOfCategory(Long categoryId) {
+    public List<Product> getAllProductsOfCategory(Long categoryId, int page, int size, boolean asc) {
+        Pageable pageable = PageRequest.of(page,size);
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() ->new AppCatchException(ErrorCode.CATEGORY_NOT_FOUND));
 
-        return productMapper.toProductResponseList(productRepository.findProductsByCategory(category));
+        if(asc){
+            return productRepository.findByShopOrderByMinPriceAsc(category.getShop(), pageable);
+        }else {
+            return productRepository.findByShopOrderByMinPriceDesc(category.getShop(), pageable);
+        }
     }
 
     //lấy tưngf sản phẩm
@@ -118,7 +187,8 @@ public class ProductService {
         return product;
     }
 
-    public Product updateProduct(Long id, ProductUpdateRequest request) {
+    @PreAuthorize("hasRole('SHOP')")
+    public Product updateProduct(Long id, ProductUpdateRequest request, List<MultipartFile> images) throws IOException {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
         User owner = userRepository.findByEmail(emailOwner).orElse(null);
@@ -130,6 +200,8 @@ public class ProductService {
                 .orElseThrow(() -> new AppCatchException(ErrorCode.PRODUCT_NOT_FOUND));
 
         productMapper.updateProduct(request, product);
+        Set<String> imagesUrl = uploadMultiImg(images, String.valueOf(owner.getId()));
+        product.setImages(imagesUrl);
 //        Event event = eventRepository.findById(request.getEventId()).orElse(null);
         Category category = categoryRepository.findById(request.getCategoryId()).orElse(null);
 
@@ -143,6 +215,7 @@ public class ProductService {
 //        product.onCreate();
         return productRepository.save(product);
     }
+    @PreAuthorize("hasRole('SHOP')")
     public void deleteProduct(Long id) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -159,6 +232,7 @@ public class ProductService {
     }
 
     //product skus
+    @PreAuthorize("hasRole('SHOP')")
     public ProductSkusResponse addProductSkusToProduct(ProductSkusCreationRequest request) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -186,6 +260,7 @@ public class ProductService {
 //    public List<ProductSkusResponse> getProductSkus(){
 //        return productSkusMapper.toProductSkusResponseList(productSkusRepository.findAll());
 //    }
+    @PreAuthorize("hasRole('SHOP')")
     public ProductSkusResponse getProductSkus(Long id) {
 
         ProductSkus productSkus = productSkusRepository.findById(id)
@@ -193,6 +268,7 @@ public class ProductService {
 
         return productSkusMapper.toProductSkusResponse(productSkus);
     }
+    @PreAuthorize("hasRole('SHOP')")
     public ProductSkusResponse updateProductSkus(Long id, ProductSkusUpdateRequest request) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -212,13 +288,26 @@ public class ProductService {
         return productSkusMapper.toProductSkusResponse(productSkusRepository.save(productSkus));
 
     }
+    @PreAuthorize("hasRole('SHOP')")
     public void deleteProductSkus(Long id) {
-        ProductSkus productSkus = productSkusRepository.findById(id)
-                        .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_SKUS_NOT_FOUND));
+        String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User owner = userRepository.findByEmail(emailOwner).orElse(null);
+
+        shopRepository.findByOwner(owner)
+                .orElseThrow(()->new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER));
+
+        ProductDetail productDetail = productDetailRepository.findById(id)
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if(!productDetail.getProduct().getShop().getOwner().equals(owner)){
+            throw new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER);
+        }
         productSkusRepository.deleteById(id);
     }
 
     //product detail
+    @PreAuthorize("hasRole('SHOP')")
     public ProductDetailResponse addProductDetailToProduct(ProductDetailCreationRequest request) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -242,6 +331,7 @@ public class ProductService {
         return productDetailResponse;
     }
 
+    @PreAuthorize("hasRole('SHOP')")
     public ProductDetailResponse getProductDetailById(Long id) {
         ProductDetail productDetail = productDetailRepository.findById(id)
                 .orElseThrow(() -> new AppCatchException(ErrorCode.PRODUCT_DETAIL_NOT_FOUND));
@@ -250,6 +340,7 @@ public class ProductService {
         productDetailResponse.setProduct(productDetail.getProduct());
         return productDetailResponse;
     }
+    @PreAuthorize("hasRole('SHOP')")
     public ProductDetailResponse updateProductDetail(Long id, ProductDetailUpdateRequest request) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -274,6 +365,7 @@ public class ProductService {
 
         return productDetailMapper.toProductDetailResponse(productDetailRepository.save(productDetail));
     }
+    @PreAuthorize("hasRole('SHOP')")
     public void deleteProductDetail(Long id) {
         String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
 
@@ -292,4 +384,97 @@ public class ProductService {
         productDetailRepository.deleteById(id);
     }
 
+    //proudct Attribute
+    @PreAuthorize("hasRole('SHOP')")
+    public ProductAttributesResponse addProductAttributes(ProductAttributesCreateRequest request) {
+        String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User owner = userRepository.findByEmail(emailOwner).orElse(null);
+
+//        shopRepository.findByOwner(owner)
+//                .orElseThrow(()->new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER));
+
+        ProductSkus productSkus = productSkusRepository.findById(request.getProductSku().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_SKUS_NOT_FOUND));
+
+        Product product = productRepository.findById(productSkus.getProduct().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if(!product.getShop().getOwner().equals(owner)){
+            throw new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER);
+        }
+
+        ProductAttributes productAttributes = productAttributesMapper.toProductAttributes(request);
+
+        Date currentDate = new Date();
+        productAttributes.setCreatedAt(currentDate);
+
+        productAttributes.setProductSku(productSkus);
+        productAttributesRepository.save(productAttributes);
+
+        return productAttributesMapper.toProductAttributesResponse(productAttributes);
+    }
+    @PreAuthorize("hasRole('SHOP')")
+    public ProductAttributesResponse getProductAttributeById(Long id) {
+        ProductAttributes productAttributes = productAttributesRepository.findById(id)
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_ATTRIBUTE_NOT_FOUND));
+
+        ProductAttributesResponse productAttributesResponse = productAttributesMapper.toProductAttributesResponse(productAttributes);
+        productAttributesResponse.setProductSku(productAttributes.getProductSku());
+        return productAttributesResponse;
+    }
+    @PreAuthorize("hasRole('SHOP')")
+    public ProductAttributesResponse updateProductAttribute(Long id, ProductAttributesUpdateRequest request) {
+        String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User owner = userRepository.findByEmail(emailOwner).orElse(null);
+
+        ProductAttributes productAttributes = productAttributesRepository.findById(id)
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_ATTRIBUTE_NOT_FOUND));
+
+        ProductSkus productSkus = productSkusRepository.findById(productAttributes.getProductSku().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_SKUS_NOT_FOUND));
+
+        Product product = productRepository.findById(productSkus.getProduct().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if(!product.getShop().getOwner().equals(owner)){
+            throw new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER);
+        }
+
+        productAttributes.setValue(request.getValue());
+        productAttributes.setType(request.getType());
+
+        Date currentDate = new Date();
+        productAttributes.setCreatedAt(currentDate);
+
+        productAttributes.setProductSku(productSkus);
+        productAttributesRepository.save(productAttributes);
+
+        ProductAttributesResponse productAttributesResponse = productAttributesMapper.toProductAttributesResponse(productAttributes);
+        productAttributesResponse.setProductSku(productAttributes.getProductSku());
+        return productAttributesResponse;
+    }
+
+    @PreAuthorize("hasRole('SHOP')")
+    public void deleteProductAttribute(Long id) {
+        String emailOwner = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User owner = userRepository.findByEmail(emailOwner).orElse(null);
+
+        ProductAttributes productAttributes = productAttributesRepository.findById(id)
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_ATTRIBUTE_NOT_FOUND));
+
+        ProductSkus productSkus = productSkusRepository.findById(productAttributes.getProductSku().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_SKUS_NOT_FOUND));
+
+        Product product = productRepository.findById(productSkus.getProduct().getId())
+                .orElseThrow(()->new AppCatchException(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if(!product.getShop().getOwner().equals(owner)){
+            throw new AppCatchException(ErrorCode.YOU_ARE_NOT_OWNER);
+        }
+
+        productAttributesRepository.deleteById(productAttributes.getId());
+    }
 }
